[AD-608] change decimalDigits to return a short
* SQLColumns returns short for decimal digits on documentation
* previously, unnecessary conversion from integer to short was done, so I've changed the SqlTypeDecimalDigits and the BinaryTypeDecimalDigits just to return shorts

+++ src/odbc/include/ignite/odbc/type_traits.h
@@ -314,7 +314,7 @@ boost::optional<int32_t> BinaryTypeNumPrecRadix(boost::optional<int16_t> type);
  * @param type SQL type.
  * @return Decimal digits.
  */
-boost::optional<int32_t> SqlTypeDecimalDigits(boost::optional<int16_t> type);
+boost::optional<int16_t> SqlTypeDecimalDigits(boost::optional<int16_t> type);
 
 /**
  * Get binary type decimal digits.
@@ -322,7 +322,7 @@ boost::optional<int32_t> SqlTypeDecimalDigits(boost::optional<int16_t> type);
  * @param type Binary type.
  * @return Decimal digits.
  */
-boost::optional<int32_t> BinaryTypeDecimalDigits(boost::optional<int16_t> type);
+boost::optional<int16_t> BinaryTypeDecimalDigits(boost::optional<int16_t> type);
 
 /**
  * Get SQL type char octet length.
